Removes old consumer from noaa package.
[#125997901]
