--- debug_printer.go
@@ -1,11 +1,0 @@
-package noaa
-
-type DebugPrinter interface {
-	Print(title, dump string)
-}
-
-type NullDebugPrinter struct {
-}
-
-func (NullDebugPrinter) Print(title, body string) {
-}
